moved BrightScriptDebugSession back out to root in order to help with the merge
